Epicalsoft.Reach.Api.Client.Net 1.3.10.41

+++ src/ES.Reach.SDK/Models/RoadType.cs
@@ -2,7 +2,7 @@
 {
     public class RoadType
     {
-        public int Code { get; set; }
+        public short Code { get; set; }
         public string Name { get; set; }
     }
 }
